Rename widgets.ts -> widget.ts
