+++ nbdime-web/src/diff/widget.ts
